Improve SocketOption API to add MembershipRequest options
I separated the get and set requirements of the SocketOption protocol because
underlying system APIs don't implement getsockopt for these membership options
and I think it's useful to encode that in the API. The changes I have made
aren't source breaking and they don't make it any more complicated to add
regular SocketOption's. In fact they make it easier with the new SimpleSocketOption
protocol.